- corrected all the forms to have PascalCase for the methods

+++ ContactlessSelfCheckout/ContactlessSelfCheckout/Product.cs
@@ -13,7 +13,7 @@ namespace ContactlessSelfCheckout
         public string productCategory;
         public double productPrice;
         public int productStock;
-        public Product(int productId, string productName, string productCategory, double productPrice, int productStock)
+        public Product(int productId, string productName, string productCategory, int productPrice, int productStock)
         {
             this.productId = productId;
             this.productName = productName;
